+++ config/autoload/api-server.global.php
@@ -45,7 +45,6 @@ return [
                 SearchDecorator\ItemDecorator::class,
                 SearchDecorator\RecipeDecorator::class,
             ],
-            ConfigKey::VERSION => '1.2.0',
         ],
     ],
 ];